+++ Backend/Api/CategoryEndpoints.cs
@@ -1,7 +1,7 @@
 ﻿using System.Data;
 using Backend.Helpers;
-using Backend.Model.Domain;
 using Backend.Types;
+using Backend.Types.Endpoint;
 using Dapper;
 using Microsoft.AspNetCore.Http.HttpResults;
