+++ Backend/Api/MonetaryFlowEndpoints.cs
@@ -1,8 +1,8 @@
 ﻿using System.Data;
 using Backend.Helpers;
 using Backend.Helpers.Cognito;
-using Backend.Model.Domain;
 using Backend.Types;
+using Backend.Types.Endpoint;
 using Dapper;
 using Microsoft.AspNetCore.Http.HttpResults;
