+++ Backend/Types/Endpoint/Category.cs
@@ -1,3 +1,3 @@
-﻿namespace Backend.Model.Domain;
+﻿namespace Backend.Types.Endpoint;
 
 public record Category(int Id, string Name);
