+++ Backend/Types/Endpoint/MonetaryFlowAdd.cs
@@ -1,0 +1,6 @@
+﻿namespace Backend.Types.Endpoint;
+
+public record MonetaryFlowAdd(
+    int GoalId,
+    int CategoryId,
+    decimal MonetaryValue);
