+++ Backend/Types/Endpoint/MonetaryFlowDelete.cs
@@ -1,0 +1,5 @@
+﻿namespace Backend.Types.Endpoint;
+
+public record MonetaryFlowDelete(
+    int FlowId
+);
