+++ Backend/Types/Endpoint/MonetaryFlowItems.cs
@@ -1,0 +1,8 @@
+﻿namespace Backend.Types.Endpoint;
+
+public record MonetaryFlowItems(
+    int FlowId,
+    string GoalName,
+    string CategoryName,
+    decimal MonetaryValue,
+    DateTime CreatedDatetime);
